+++ Library/include/actuators/Light.h
@@ -20,7 +20,7 @@
 //  Stonefish
 //
 //  Created by Patryk Cieslak on 4/7/17.
-//  Copyright (c) 2017-2020 Patryk Cieslak. All rights reserved.
+//  Copyright (c) 2017-2024 Patryk Cieslak. All rights reserved.
 //
 
 #ifndef __Stonefish_Light__
@@ -105,6 +105,7 @@ namespace sf
         
     private:
         void InitGraphics();
+        
         //attach -> SolidEntity
 		StaticEntity* attach2;
         AnimatedEntity* attach3;