+++ docs/changelog.rst
@@ -13,6 +13,7 @@ The changelog of the library code is presented below. **Breaking changes** were 
 -  Implemented new trajectory generator for animated bodies utilising B-splines (now default)
 -  Fixed IMU readings, adding the missing gravitational and centrifugal accelerations
 -  Extended glue to support joining links of two robots together
+-  Added a watchdog timer to the actuators, including parser support
 -  Added access to the viscous and quadratic hydrodynamic drag coefficients, including parser support
 -  Added an option to set internal parts of a compound body as always visible
 -  Added access to the computed wetted surface area and submerged volume